PDF Phase 0 (#734): move viewrect codec to core; export + harden content Part of #351. Completes the @semiont/content text-layer slice and the package-boundary refactor #734 requires before Phase 2 (#736). core: move PdfCoordinate + the RFC 3778 viewrect FragmentSelector codec (createFragmentSelector / parseFragmentSelector / getPageFromFragment) out of react-ui into @semiont/core, so server packages (jobs, make-meaning) and the browser all import it without reaching into the UI. Drop the dead try/catch in parseFragmentSelector (no statement in it can throw). react-ui: pdf-coordinates.ts keeps only the canvas-pixel transforms and imports PdfCoordinate from core; PdfAnnotationCanvas and the lib tests updated to the new import sources. content: export PdfTextLayer / PdfTextItem / PdfPageInfo and extractPdfTextLayer / locate from the package index; source PdfCoordinate from core; tidy pdf-text-layer.ts JSDoc. extractPdfTextLayer now passes isEvalSupported: false — no eval of code derived from an untrusted PDF (CVE-2024-4367); getTextContent needs no codegen, so it costs nothing. tests: move the codec property tests to core; add the #734 round-trip spike and a locate() -> core-codec serialization round-trip (the handoff #736 depends on).
