+++ packages/content/src/pdf-text-layer.ts
@@ -1,43 +1,25 @@
 /**
-* PDF Text Layer Types
-*
-* Represents the extracted text layer from a native PDF, 
-* including per-character geometry in PDF point coordinates 
-* originating from the bottom left.
-* 
-* Note: PdfCoordinate is defined here temporarily. It will move to
-* @semiont/core alongside the `FragmentSelector` before Phase 2 (#736),
-* so that packages/jobs can serialize viewrects without depending on react-ui.
-* Update the import when that refactor lands.
-* 
-*/
-
+ * PDF Text Layer Types
+ *
+ * Represents the extracted text layer from a native PDF, including per-character
+ * geometry in PDF point coordinates originating from the bottom-left of the page
+ * (Y increases upward). The Y-flip to canvas pixels happens downstream in the
+ * browser; the server has no canvas.
+ *
+ * `PdfCoordinate` — the geometry `locate()` emits — lives in `@semiont/core`
+ * alongside the viewrect FragmentSelector codec.
+ */
 
 /**
-* A bounding rectangle in PDF point coordinates
-* Originate from bottom-left of the page; Y increases upward.
-* Y-flip to canvas pixels happens elsewhere.
-* 
-* This should be moved to @semiont/core before Phase 2.
-*/
-export interface PdfCoordinate {
-    page: number;
-    x: number;
-    y: number;
-    width: number;
-    height: number;
-}
-
-/**
-* A single text item from the PDF text layer.
-* Character offsets refer to positions in `PdfTextLayer.text`.
-*/
+ * A single text item from the PDF text layer.
+ * Character offsets refer to positions in `PdfTextLayer.text`.
+ */
 export interface PdfTextItem {
     start: number;  // Char offset in `PdfTextLayer.text` (inclusive)
-    end: number;  // Char offset in `PdfTextLayer.text` (exclusive)
-    page: number;  // 1-indexed page number
-    x: number;  // X position in PDF points (origin: bottom-left of page) 
-    y: number;  // Y position in PDF points (origin: bottom-left of page)
+    end: number;    // Char offset in `PdfTextLayer.text` (exclusive)
+    page: number;   // 1-indexed page number
+    x: number;      // X position in PDF points (origin: bottom-left of page)
+    y: number;      // Y position in PDF points (origin: bottom-left of page)
     width: number;
     height: number;
 }
@@ -52,8 +34,8 @@ export interface PdfPageInfo {
 /**
  * The full extracted text layer for a PDF.
  * `text` is the reading-order concatenation across all pages.
- * `items` carry per-character ranges into `text` plus PDF-point geometry
-*/
+ * `items` carry per-character ranges into `text` plus PDF-point geometry.
+ */
 export interface PdfTextLayer {
     pages: PdfPageInfo[];
     text: string;
